+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Optimizing `#.mergeEdge`.
 * Fixing `#.edges(A, B)` with directed self loops.
 * Fixing case when `nodeAdded` events would not be fired when using `#.emptyCopy`.
+* Correctly emitting `nodeAttributesUpdated` event when node is merged.
 
 ## 0.18.0